# server/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Room:
    properties:
      attendees:
        items:
          $ref: '#/definitions/model.User'
        type: array
      consolidated:
        type: string
      createdAt:
        type: string
      date:
        type: string
      description:
        type: string
      host:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Associations
      hostId:
        type: integer
      id:
        type: string
      imageUrl:
        type: string
      isClosed:
        type: boolean
      isPrivate:
        type: boolean
      name:
        type: string
      noOfAttendees:
        type: integer
      time:
        type: string
      updatedAt:
        type: string
      venue:
        type: string
      venuePlaceId:
        type: string
      venueUrl:
        type: string
    type: object
...
  request.CreateRoomRequest:
    properties:
      invitees:
        items:
          type: string
        type: array
      room:
        $ref: '#/definitions/model.Room'
    type: object
...
  response.CreateRoomResponse:
    properties:
      invites:
        items:
          $ref: '#/definitions/model.RoomInvite'
        type: array
      room:
        $ref: '#/definitions/model.Room'
    type: object
...
  response.RoomDto:
    properties:
      attendees:
        items:
          $ref: '#/definitions/response.AttendeesDto'
        type: array
      consolidated:
        type: string
      date:
        type: string
      description:
        type: string
      host:
        $ref: '#/definitions/response.AttendeesDto'
      id:
        type: string
      imageUrl:
        type: string
      isClosed:
        type: boolean
      isPrivate:
        type: boolean
      name:
        type: string
      noOfAttendees:
        type: integer
      time:
        type: string
      venue:
        type: string
      venueUrl:
        type: string
    type: object
  utils.EmptyApiResponse:
    properties:
      data:
        type: object
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API server for JustJio.
  title: JustJio API
  version: "1.0"
paths:
...
  /rooms:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of rooms for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved rooms successfully
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.Room'
                type: array
              message:
                type: string
              status:
                type: string
            type: object
        "404":
          description: No rooms found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Get user's rooms
      tags:
      - Rooms
    post:
      consumes:
      - application/json
      description: Creates a new room and optionally sends invitations to specified
        users
      parameters:
      - description: Room creation details
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/request.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created room successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/response.CreateRoomResponse'
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Failed to create room and invites
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Create room
      tags:
      - Rooms
  /rooms/{roomId}:
    get:
      consumes:
      - application/json
      description: Retrieves details of a specific room by its ID
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved room successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/response.RoomDto'
              message:
                type: string
              status:
                type: string
            type: object
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Get room details
      tags:
      - Rooms
    patch:
      consumes:
      - application/json
      description: Updates details of an existing room (host only)
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Room update details
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/request.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Edited room successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/model.Room'
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "401":
          description: Only hosts can edit rooms
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Edit room
      tags:
      - Rooms
  /rooms/{roomId}/attendees:
    get:
      consumes:
      - application/json
      description: Retrieves the list of attendees for a specific room
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved room attendees successfully
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.User'
                type: array
              message:
                type: string
              status:
                type: string
            type: object
        "404":
          description: No attendees found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      summary: Get room attendees
      tags:
      - Rooms
  /rooms/{roomId}/close:
    patch:
      consumes:
      - application/json
      description: Closes a room (host only, no unconsolidated bills allowed)
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Closed room successfully
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "401":
          description: Only hosts are allowed to close rooms
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "409":
          description: Cannot close room with unconsolidated bills
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Close room
      tags:
      - Rooms
  /rooms/{roomId}/invitations:
    post:
      consumes:
      - application/json
      description: Invites multiple users to a room (host only)
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: User invitation details
        in: body
        name: invites
        required: true
        schema:
          $ref: '#/definitions/request.InviteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invited users successfully
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.RoomInvite'
                type: array
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "401":
          description: Only hosts are allowed to invite users
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: Room / User not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "409":
          description: User is already in the room or already has pending invite
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Invite users to room
      tags:
      - Room Invitations
  /rooms/{roomId}/invitations/respond:
    patch:
      consumes:
      - application/json
      description: Accepts or rejects a room invitation
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Invitation response
        in: body
        name: response
        required: true
        schema:
          $ref: '#/definitions/request.RespondToRoomInviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rejected room invitation successfully
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Respond to room invitation
      tags:
      - Room Invitations
  /rooms/{roomId}/join:
    post:
      consumes:
      - application/json
      description: Allows a user to join a public room
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Joined room successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/response.JoinRoomResponse'
              message:
                type: string
              status:
                type: string
            type: object
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "409":
          description: User is already in room
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Join room
      tags:
      - Rooms
  /rooms/{roomId}/leave:
    delete:
      consumes:
      - application/json
      description: Allows a user to leave a room (not allowed for hosts or with unconsolidated
        bills)
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Left room successfully
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "409":
          description: Cannot leave room with unconsolidated bills or host cannot
            leave room
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Leave room
      tags:
      - Rooms
...
  /rooms/{roomId}/uninvited-friends:
    get:
      consumes:
      - application/json
      description: Retrieves the list of friends who haven't been invited to the specified
        room
      parameters:
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved uninvited friends successfully
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.User'
                type: array
              message:
                type: string
              status:
                type: string
            type: object
        "404":
          description: No uninvited friends found
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.EmptyApiResponse'
      security:
      - BearerAuth: []
      summary: Get uninvited friends for room
      tags:
      - Rooms
...
schemes:
- http
- https
swagger: "2.0"